+++ src/cli/ProxyCertificateDelegator.h
@@ -55,7 +55,7 @@ public:
 	 * @param userRequestedDelegationExpTime - user specified delegation expire time
 	 * 											(0 means that the default value should be used)
 	 */
-	ProxyCertificateDelegator(string endpoint, string delegationId, int userRequestedDelegationExpTime, MsgPrinter& printer);
+	ProxyCertificateDelegator(string endpoint, string delegationId, long userRequestedDelegationExpTime, MsgPrinter& printer);
 
 	/**
 	 * Destructor
@@ -97,7 +97,7 @@ private:
 	string endpoint;
 
 	/// user defined proxy certificate expiration time
-	int userRequestedDelegationExpTime;
+	long userRequestedDelegationExpTime;
 
 	/// delegation context (a facade for the GSoap delegation client)
 	glite_delegation_ctx *dctx;
